Add helpdesk feature for employee

+++ controllers/helpdeskEmployee.go
@@ -12,7 +12,8 @@ import (
 	"time"
 )
 
-func CreateHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
+// CreateHelpdeskByEmployee memungkinkan karyawan untuk menambahkan data helpdesk
+func CreateHelpdeskByEmployee(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 	return func(c echo.Context) error {
 		// Extract and verify the JWT token
 		tokenString := c.Request().Header.Get("Authorization")
@@ -35,17 +36,12 @@ func CreateHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusUnauthorized, errorResponse)
 		}
 
-		// Check if the user is an admin
-		var adminUser models.Admin
-		result := db.Where("username = ?", username).First(&adminUser)
+		// Retrieve the employee based on the username
+		var employee models.Employee
+		result := db.Where("username = ?", username).First(&employee)
 		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Admin user not found"}
+			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
 			return c.JSON(http.StatusNotFound, errorResponse)
-		}
-
-		if !adminUser.IsAdminHR {
-			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Access denied"}
-			return c.JSON(http.StatusForbidden, errorResponse)
 		}
 
 		// Bind the helpdesk data from the request body
@@ -56,10 +52,15 @@ func CreateHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 		}
 
 		// Validate helpdesk data
-		if helpdesk.Subject == "" || helpdesk.Priority == "" || helpdesk.Description == "" || helpdesk.DepartmentID == 0 || helpdesk.EmployeeID == 0 {
+		if helpdesk.Subject == "" || helpdesk.Priority == "" || helpdesk.Description == "" || helpdesk.DepartmentID == 0 {
 			errorResponse := helper.Response{Code: http.StatusBadRequest, Error: true, Message: "All fields are required"}
 			return c.JSON(http.StatusBadRequest, errorResponse)
 		}
+
+		// Set the employee ID and username
+		helpdesk.EmployeeID = employee.ID
+		helpdesk.EmployeeUsername = employee.Username
+		helpdesk.EmployeeFullName = employee.FirstName + " " + employee.LastName
 
 		// Check if the department with the given ID exists
 		var existingDepartment models.Department
@@ -69,18 +70,8 @@ func CreateHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
-		// Check if the employee with the given ID exists
-		var existingEmployee models.Employee
-		result = db.First(&existingEmployee, helpdesk.EmployeeID)
-		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
-			return c.JSON(http.StatusNotFound, errorResponse)
-		}
-
-		// Set the DepartmentName and EmployeeUsername
+		// Set the DepartmentName
 		helpdesk.DepartmentName = existingDepartment.DepartmentName
-		helpdesk.EmployeeUsername = existingEmployee.Username
-		helpdesk.EmployeeFullName = existingEmployee.FirstName + " " + existingEmployee.LastName
 
 		// Set the created timestamp
 		currentTime := time.Now()
@@ -100,7 +91,8 @@ func CreateHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 	}
 }
 
-func GetAllHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
+// GetHelpdeskByEmployee memungkinkan karyawan untuk melihat semua data helpdesk miliknya
+func GetAllHelpdeskByEmployee(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 	return func(c echo.Context) error {
 		// Extract and verify the JWT token
 		tokenString := c.Request().Header.Get("Authorization")
@@ -123,35 +115,31 @@ func GetAllHelpdeskByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusUnauthorized, errorResponse)
 		}
 
-		// Check if the user is an admin
-		var adminUser models.Admin
-		result := db.Where("username = ?", username).First(&adminUser)
+		// Retrieve the employee based on the username
+		var employee models.Employee
+		result := db.Where("username = ?", username).First(&employee)
 		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Admin user not found"}
+			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
 			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
-		if !adminUser.IsAdminHR {
-			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Access denied"}
-			return c.JSON(http.StatusForbidden, errorResponse)
-		}
-
-		// Retrieve all helpdesk data from the database
-		var helpdesks []models.Helpdesk
-		db.Find(&helpdesks)
+		// Retrieve all helpdesk data for the employee
+		var helpdeskList []models.Helpdesk
+		db.Where("employee_id = ?", employee.ID).Find(&helpdeskList)
 
 		// Respond with success
-		successResponse := helper.Response{
-			Code:      http.StatusOK,
-			Error:     false,
-			Message:   "All helpdesk data retrieved successfully",
-			Helpdesks: helpdesks,
+		successResponse := map[string]interface{}{
+			"code":     http.StatusOK,
+			"error":    false,
+			"message":  "Helpdesk data retrieved successfully",
+			"helpdesk": helpdeskList,
 		}
 		return c.JSON(http.StatusOK, successResponse)
 	}
 }
 
-func GetHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
+// GetHelpdeskByIDForEmployee memungkinkan karyawan untuk melihat data helpdesk miliknya berdasarkan ID helpdesk
+func GetHelpdeskByIDByEmployee(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 	return func(c echo.Context) error {
 		// Extract and verify the JWT token
 		tokenString := c.Request().Header.Get("Authorization")
@@ -174,17 +162,79 @@ func GetHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusUnauthorized, errorResponse)
 		}
 
-		// Check if the user is an admin
-		var adminUser models.Admin
-		result := db.Where("username = ?", username).First(&adminUser)
+		// Retrieve the employee based on the username
+		var employee models.Employee
+		result := db.Where("username = ?", username).First(&employee)
 		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Admin user not found"}
+			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
 			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
-		if !adminUser.IsAdminHR {
-			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Access denied"}
+		// Extract ID parameter from the request
+		id := c.Param("id")
+
+		// Parse ID to uint
+		helpdeskID, err := strconv.ParseUint(id, 10, 64)
+		if err != nil {
+			errorResponse := helper.Response{Code: http.StatusBadRequest, Error: true, Message: "Invalid ID format"}
+			return c.JSON(http.StatusBadRequest, errorResponse)
+		}
+
+		// Retrieve the helpdesk data from the database by ID
+		var helpdesk models.Helpdesk
+		result = db.First(&helpdesk, helpdeskID)
+		if result.Error != nil {
+			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Helpdesk not found"}
+			return c.JSON(http.StatusNotFound, errorResponse)
+		}
+
+		// Check if the helpdesk belongs to the employee
+		if helpdesk.EmployeeID != employee.ID {
+			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Helpdesk does not belong to the employee"}
 			return c.JSON(http.StatusForbidden, errorResponse)
+		}
+
+		// Respond with success
+		successResponse := map[string]interface{}{
+			"code":     http.StatusOK,
+			"error":    false,
+			"message":  "Helpdesk data retrieved successfully",
+			"helpdesk": helpdesk,
+		}
+		return c.JSON(http.StatusOK, successResponse)
+	}
+}
+
+// UpdateHelpdeskByIDForEmployee memungkinkan karyawan untuk mengedit data helpdesk miliknya berdasarkan ID helpdesk
+func UpdateHelpdeskByIDByEmployee(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
+	return func(c echo.Context) error {
+		// Extract and verify the JWT token
+		tokenString := c.Request().Header.Get("Authorization")
+		if tokenString == "" {
+			errorResponse := helper.Response{Code: http.StatusUnauthorized, Error: true, Message: "Authorization token is missing"}
+			return c.JSON(http.StatusUnauthorized, errorResponse)
+		}
+
+		authParts := strings.SplitN(tokenString, " ", 2)
+		if len(authParts) != 2 || authParts[0] != "Bearer" {
+			errorResponse := helper.Response{Code: http.StatusUnauthorized, Error: true, Message: "Invalid token format"}
+			return c.JSON(http.StatusUnauthorized, errorResponse)
+		}
+
+		tokenString = authParts[1]
+
+		username, err := middleware.VerifyToken(tokenString, secretKey)
+		if err != nil {
+			errorResponse := helper.Response{Code: http.StatusUnauthorized, Error: true, Message: "Invalid token"}
+			return c.JSON(http.StatusUnauthorized, errorResponse)
+		}
+
+		// Retrieve the employee based on the username
+		var employee models.Employee
+		result := db.Where("username = ?", username).First(&employee)
+		if result.Error != nil {
+			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
+			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
 		// Extract helpdesk ID from the request params
@@ -203,67 +253,10 @@ func GetHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
-		// Respond with success
-		successResponse := helper.Response{
-			Code:     http.StatusOK,
-			Error:    false,
-			Message:  "Helpdesk data retrieved successfully",
-			Helpdesk: &helpdesk,
-		}
-		return c.JSON(http.StatusOK, successResponse)
-	}
-}
-
-func UpdateHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
-	return func(c echo.Context) error {
-		// Extract and verify the JWT token
-		tokenString := c.Request().Header.Get("Authorization")
-		if tokenString == "" {
-			errorResponse := helper.Response{Code: http.StatusUnauthorized, Error: true, Message: "Authorization token is missing"}
-			return c.JSON(http.StatusUnauthorized, errorResponse)
-		}
-
-		authParts := strings.SplitN(tokenString, " ", 2)
-		if len(authParts) != 2 || authParts[0] != "Bearer" {
-			errorResponse := helper.Response{Code: http.StatusUnauthorized, Error: true, Message: "Invalid token format"}
-			return c.JSON(http.StatusUnauthorized, errorResponse)
-		}
-
-		tokenString = authParts[1]
-
-		username, err := middleware.VerifyToken(tokenString, secretKey)
-		if err != nil {
-			errorResponse := helper.Response{Code: http.StatusUnauthorized, Error: true, Message: "Invalid token"}
-			return c.JSON(http.StatusUnauthorized, errorResponse)
-		}
-
-		// Check if the user is an admin
-		var adminUser models.Admin
-		result := db.Where("username = ?", username).First(&adminUser)
-		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Admin user not found"}
-			return c.JSON(http.StatusNotFound, errorResponse)
-		}
-
-		if !adminUser.IsAdminHR {
-			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Access denied"}
+		// Check if the helpdesk belongs to the employee
+		if helpdesk.EmployeeID != employee.ID {
+			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Helpdesk does not belong to the employee"}
 			return c.JSON(http.StatusForbidden, errorResponse)
-		}
-
-		// Extract helpdesk ID from the request params
-		helpdeskIDStr := c.Param("id")
-		helpdeskID, err := strconv.ParseUint(helpdeskIDStr, 10, 32)
-		if err != nil {
-			errorResponse := helper.Response{Code: http.StatusBadRequest, Error: true, Message: "Invalid helpdesk ID format"}
-			return c.JSON(http.StatusBadRequest, errorResponse)
-		}
-
-		// Retrieve helpdesk data from the database based on ID
-		var helpdesk models.Helpdesk
-		result = db.First(&helpdesk, uint(helpdeskID))
-		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Helpdesk data not found"}
-			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
 		// Bind updated helpdesk data from the request body
@@ -276,10 +269,6 @@ func UpdateHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 		// Update helpdesk data selectively
 		if updatedHelpdesk.Subject != "" {
 			helpdesk.Subject = updatedHelpdesk.Subject
-		}
-
-		if updatedHelpdesk.Priority != "" {
-			helpdesk.Priority = updatedHelpdesk.Priority
 		}
 
 		if updatedHelpdesk.DepartmentID != 0 {
@@ -295,26 +284,12 @@ func UpdateHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			helpdesk.DepartmentName = existingDepartment.DepartmentName
 		}
 
-		if updatedHelpdesk.EmployeeID != 0 {
-			// Check if the employee with the given ID exists
-			var existingEmployee models.Employee
-			result = db.First(&existingEmployee, updatedHelpdesk.EmployeeID)
-			if result.Error != nil {
-				errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
-				return c.JSON(http.StatusNotFound, errorResponse)
-			}
-			// Set the EmployeeUsername
-			helpdesk.EmployeeID = updatedHelpdesk.EmployeeID
-			helpdesk.EmployeeUsername = existingEmployee.Username
-			helpdesk.EmployeeFullName = existingEmployee.FirstName + " " + existingEmployee.LastName
+		if updatedHelpdesk.Priority != "" {
+			helpdesk.Priority = updatedHelpdesk.Priority
 		}
 
 		if updatedHelpdesk.Description != "" {
 			helpdesk.Description = updatedHelpdesk.Description
-		}
-
-		if updatedHelpdesk.TicketStatus != "" {
-			helpdesk.TicketStatus = updatedHelpdesk.TicketStatus
 		}
 
 		// Set the updated timestamp
@@ -334,7 +309,8 @@ func UpdateHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 	}
 }
 
-func DeleteHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
+// DeleteHelpdeskByIDForEmployee memungkinkan karyawan untuk menghapus data helpdesk miliknya berdasarkan ID helpdesk
+func DeleteHelpdeskByIDByEmployee(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 	return func(c echo.Context) error {
 		// Extract and verify the JWT token
 		tokenString := c.Request().Header.Get("Authorization")
@@ -357,17 +333,12 @@ func DeleteHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusUnauthorized, errorResponse)
 		}
 
-		// Check if the user is an admin
-		var adminUser models.Admin
-		result := db.Where("username = ?", username).First(&adminUser)
+		// Retrieve the employee based on the username
+		var employee models.Employee
+		result := db.Where("username = ?", username).First(&employee)
 		if result.Error != nil {
-			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Admin user not found"}
+			errorResponse := helper.Response{Code: http.StatusNotFound, Error: true, Message: "Employee not found"}
 			return c.JSON(http.StatusNotFound, errorResponse)
-		}
-
-		if !adminUser.IsAdminHR {
-			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Access denied"}
-			return c.JSON(http.StatusForbidden, errorResponse)
 		}
 
 		// Extract helpdesk ID from the request params
@@ -386,14 +357,20 @@ func DeleteHelpdeskByIDByAdmin(db *gorm.DB, secretKey []byte) echo.HandlerFunc {
 			return c.JSON(http.StatusNotFound, errorResponse)
 		}
 
-		// Delete helpdesk data from the database
+		// Check if the helpdesk belongs to the employee
+		if helpdesk.EmployeeID != employee.ID {
+			errorResponse := helper.Response{Code: http.StatusForbidden, Error: true, Message: "Helpdesk does not belong to the employee"}
+			return c.JSON(http.StatusForbidden, errorResponse)
+		}
+
+		// Delete the helpdesk from the database
 		db.Delete(&helpdesk)
 
 		// Respond with success
 		successResponse := helper.Response{
 			Code:    http.StatusOK,
 			Error:   false,
-			Message: "Helpdesk data deleted successfully",
+			Message: "Helpdesk deleted successfully",
 		}
 		return c.JSON(http.StatusOK, successResponse)
 	}
